+++ PL0/src/main/java/parser/statements/ReadStatement.java
@@ -24,7 +24,7 @@ public class ReadStatement implements IStatement {
                 Variables.put(identifier, new StringValue(newValue));
             }
             else {
-                Variables.put(identifier, new NumberValue(Double.parseDouble(newValue)));
+                Variables.put(identifier, new NumberValue(Integer.parseInt(newValue)));
             }
         } catch (NumberFormatException e) {
             System.err.println("Datatype " + value.getClass().getSimpleName() + " doesn't match with '" + newValue + "' value.");
